+++ lib/generators/leva/eval_generator.rb
@@ -3,10 +3,10 @@
 module Leva
   module Generators
     class EvalGenerator < Rails::Generators::NamedBase
-      source_root File.expand_path('templates', __dir__)
+      source_root File.expand_path("templates", __dir__)
 
       def create_eval_file
-        template 'eval.rb.erb', File.join('app/evals', class_path, "#{file_name}_eval.rb")
+        template "eval.rb.erb", File.join("app/evals", class_path, "#{file_name}_eval.rb")
       end
 
       private
@@ -16,7 +16,7 @@ module Leva
       end
 
       def remove_possible_suffix(name)
-        name.sub(/_?eval$/i, '')
+        name.sub(/_?eval$/i, "")
       end
     end
   end